% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{build_preference}
\alias{build_preference}
\title{Build a preference}
\usage{
build_preference(
  back_urls,
  items,
  notification_url = NULL,
  external_reference = NULL,
  integrator_id = NULL
)
}
\arguments{
\item{back_urls}{List with 3 named elements \code{success}, \code{failure} and \code{pending}.}

\item{items}{A data frame with one row per product item.}
}
\description{
A payment preference is an object or set of of information that represents the product or service you want to charge for. Within the Mercado Pago ecosystem, this object is known as \code{preference}. (https://www.mercadopago.com.pe/developers/en/docs/checkout-pro/configure-development-enviroment).
}
