% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{mercadopago}
\alias{mercadopago}
\title{Main python module}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 1.
}
\value{
the mercadopago Python module
}
\description{
Main python module
}
\keyword{datasets}
