% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{request_options}
\alias{request_options}
\title{Request options}
\usage{
request_options(
  access_token = NULL,
  connection_timeout = 60L,
  custom_headers = NULL,
  corporation_id = NULL,
  integrator_id = NULL,
  platform_id = NULL,
  max_retries = 3L
)
}
\description{
Request options
}
